+++ src/com/company/Puzzle.java
@@ -6,7 +6,7 @@ abstract class Puzzle {
     public Puzzle(List<String> lines) {
         this.lines = lines;
     }
-    public abstract long runIt();
+    public abstract int runIt();
 
     protected List<String> lines;
 }
